refactor: 💡 user,navbar,tradetab,opensubtab,historysubtab,
position sub tab, ordersection mobile version combined with desktop one
