+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.8.0+30.8",
+  "version": "0.8.0+33.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
