+++ src/components/open_sub_tab/open_sub_tab.tsx
@@ -6,10 +6,13 @@ import {MarketContext} from '../../contexts/market_context';
 import {getStateCode, toDisplayCFDOrder} from '../../lib/common';
 import {IDisplayCFDOrder} from '../../interfaces/tidebit_defi_background/display_accepted_cfd_order';
 import {SKELETON_DISPLAY_TIME} from '../../constants/display';
+import {useGlobal} from '../../contexts/global_context';
+import {LayoutAssertion} from '../../constants/layout_assertion';
 
 const OpenSubTab = () => {
   const userCtx = useContext(UserContext);
   const marketCtx = useContext(MarketContext);
+  const globalCtx = useGlobal();
 
   const [isLoading, setIsLoading] = useState(true);
   const [cfds, setCfds] = useState<IDisplayCFDOrder[]>([]);
@@ -61,12 +64,23 @@ const OpenSubTab = () => {
     })
   );
 
+  const desktopLayout = (
+    <div className="h-full overflow-y-auto overflow-x-hidden pb-40">{openPositionList}</div>
+  );
+
+  const mobileLayout = (
+    <div className="flex w-screen flex-col overflow-x-hidden px-8 sm:w-400px">
+      <div className="h-80vh overflow-y-auto px-4">{openPositionList}</div>
+    </div>
+  );
+
+  const displayedLayout =
+    globalCtx.layoutAssertion === LayoutAssertion.MOBILE ? mobileLayout : desktopLayout;
+
   return (
-    <>
-      <SkeletonTheme baseColor="#202020" highlightColor="#444">
-        <div className="h-full overflow-y-auto overflow-x-hidden pb-40">{openPositionList}</div>
-      </SkeletonTheme>
-    </>
+    <SkeletonTheme baseColor="#202020" highlightColor="#444">
+      {displayedLayout}{' '}
+    </SkeletonTheme>
   );
 };
