+++ src/components/order_section/order_section.tsx
@@ -7,29 +7,64 @@ import {UserContext} from '../../contexts/user_context';
 import {MarketContext} from '../../contexts/market_context';
 import PositionVisitorTab from '../position_visitor_tab/position_visitor_tab';
 import {useTranslation} from 'next-i18next';
+import {WalletConnectButton} from '../wallet_connect_button/wallet_connect_button';
+import {useGlobal} from '../../contexts/global_context';
+import {LayoutAssertion} from '../../constants/layout_assertion';
 
 type TranslateFunction = (s: string) => string;
+
+export enum POSITION_TAB {
+  OPEN = 'OPEN',
+  HISTORY = 'HISTORY',
+}
+
+export enum ORDER_SECTION_TAB {
+  TRADE = 'TRADE',
+  POSITION = 'POSITION',
+}
 
 const OrderSection = () => {
   const {t}: {t: TranslateFunction} = useTranslation('common');
 
+  const globalCtx = useGlobal();
   const userCtx = useContext(UserContext);
   const {isCFDTradable} = useContext(MarketContext);
 
   const [activeTab, setActiveTab] = useState('Trade');
 
+  const [positionActiveTab, setPositionActiveTab] = useState(POSITION_TAB.OPEN);
+  const [showPositionMenu, setShowPositionMenu] = useState(false);
+
+  const subMenuClickHandler = () => {
+    setShowPositionMenu(!showPositionMenu);
+  };
+
+  const openTabClickHandler = () => {
+    setPositionActiveTab(POSITION_TAB.OPEN);
+  };
+
+  const historyTabClickHandler = () => {
+    setPositionActiveTab(POSITION_TAB.HISTORY);
+  };
+
   const tradeTabClickHandler = () => {
-    setActiveTab('Trade');
+    setActiveTab(ORDER_SECTION_TAB.TRADE);
   };
 
   const positionTabClickHandler = () => {
-    setActiveTab('Position');
+    setActiveTab(ORDER_SECTION_TAB.POSITION);
   };
 
   const currentTab =
-    activeTab === 'Position' ? (
+    activeTab === ORDER_SECTION_TAB.POSITION ? (
       userCtx.enableServiceTerm ? (
-        <PositionTab />
+        <PositionTab
+          showSubMenu={showPositionMenu}
+          subMenuClickHandler={subMenuClickHandler}
+          activePositionTabMobile={positionActiveTab}
+          openTabClickHandler={openTabClickHandler}
+          historyTabClickHandler={historyTabClickHandler}
+        />
       ) : (
         <PositionVisitorTab />
       )
@@ -41,11 +76,23 @@ const OrderSection = () => {
       <TradeTab />
     );
 
-  const activeTradeTabStyle =
-    activeTab == 'Trade' ? 'bg-darkGray7 text-lightWhite' : 'bg-darkGray6 text-lightGray';
+  const tradeTabStyle =
+    activeTab == ORDER_SECTION_TAB.TRADE
+      ? 'bg-darkGray7 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
+  const positionTabStyle =
+    activeTab == ORDER_SECTION_TAB.POSITION
+      ? 'bg-darkGray7 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
 
-  const activePositionTabStyle =
-    activeTab == 'Position' ? 'bg-darkGray7 text-lightWhite' : 'bg-darkGray6 text-lightGray';
+  const openTabStyle =
+    positionActiveTab == POSITION_TAB.OPEN
+      ? 'bg-darkGray8 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
+  const historyTabStyle =
+    positionActiveTab == POSITION_TAB.HISTORY
+      ? 'bg-darkGray8 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
 
   const tabPart = (
     <>
@@ -53,7 +100,7 @@ const OrderSection = () => {
         <div className="mr-3px w-full">
           <button
             type="button"
-            className={`${activeTradeTabStyle} inline-block w-full rounded-t-2xl px-40px py-2 hover:cursor-pointer`}
+            className={`${tradeTabStyle} inline-block w-full rounded-t-2xl px-40px py-2 hover:cursor-pointer`}
             onClick={tradeTabClickHandler}
           >
             {t('TRADE_PAGE.TRADE_TAB')}
@@ -62,7 +109,7 @@ const OrderSection = () => {
         <div className="w-full">
           <button
             type="button"
-            className={`${activePositionTabStyle} inline-block w-full rounded-t-2xl px-40px py-2 hover:cursor-pointer`}
+            className={`${positionTabStyle} inline-block w-full rounded-t-2xl px-40px py-2 hover:cursor-pointer`}
             onClick={positionTabClickHandler}
           >
             {t('TRADE_PAGE.POSITION_TAB')}
@@ -72,17 +119,75 @@ const OrderSection = () => {
     </>
   );
 
-  const displayedTab = (
-    <>
-      {/* tab section */}
-      <div className="fixed top-70px right-0">{tabPart}</div>
+  const displayedPositionSubTab = showPositionMenu ? (
+    <div
+      className={`flex items-center ${'h-76px w-280px'} bg-darkGray transition-all duration-300`}
+    >
+      <ul className="ml-5 flex basis-full items-center text-center text-sm font-medium">
+        <li className="w-full">
+          <button
+            onClick={openTabClickHandler}
+            className={`${openTabStyle} inline-block w-full rounded-md px-7 py-3`}
+          >
+            {t('TRADE_PAGE.POSITION_TAB_OPEN')}
+          </button>
+        </li>
+        <li className="ml-1 w-full">
+          <button
+            onClick={historyTabClickHandler}
+            className={`${historyTabStyle} inline-block w-full rounded-md px-7 py-3`}
+          >
+            {t('TRADE_PAGE.POSITION_TAB_HISTORY')}
+          </button>
+        </li>
+      </ul>
+    </div>
+  ) : (
+    <TradeTab />
+  );
 
-      {/* trade or position section */}
+  const displayedMobileTab = userCtx.enableServiceTerm ? (
+    <div className="flex items-center">
+      {displayedPositionSubTab}
+      <div className="ml-4">
+        <PositionTab
+          showSubMenu={showPositionMenu}
+          subMenuClickHandler={subMenuClickHandler}
+          activePositionTabMobile={positionActiveTab}
+          openTabClickHandler={openTabClickHandler}
+          historyTabClickHandler={historyTabClickHandler}
+        />
+      </div>
+    </div>
+  ) : (
+    <div className="flex items-center justify-center">
+      <div className="mr-10 text-center text-sm text-lightGray">
+        {t('TRADE_PAGE.WALLET_CONNECT_DESCRIPTION_MOBILE')}
+      </div>
+
+      <WalletConnectButton className="px-5 py-2" />
+    </div>
+  );
+
+  const desktopLayout = (
+    <>
+      <div className="fixed top-70px right-0">{tabPart}</div>
       {currentTab}
     </>
   );
 
-  return <>{displayedTab}</>;
+  const mobileLayout = (
+    <div
+      className={`z-10 ${'h-76px'} fixed bottom-0 flex w-screen basis-full items-center justify-center bg-darkGray px-5 py-3`}
+    >
+      {displayedMobileTab}
+    </div>
+  );
+
+  const displayedLayout =
+    globalCtx.layoutAssertion === LayoutAssertion.MOBILE ? mobileLayout : desktopLayout;
+
+  return <>{displayedLayout}</>;
 };
 
 export default OrderSection;
