+++ src/components/position_tab/position_tab.tsx
@@ -1,29 +1,67 @@
-import React, {useState} from 'react';
+import React, {Dispatch, SetStateAction, useState} from 'react';
 import OpenSubTab from '../open_sub_tab/open_sub_tab';
 import HistorySubTab from '../history_sub_tab/history_sub_tab';
 import {useTranslation} from 'next-i18next';
+import {useGlobal} from '../../contexts/global_context';
+import {LayoutAssertion} from '../../constants/layout_assertion';
+import {POSITION_TAB} from '../order_section/order_section';
+import {ImCross} from 'react-icons/im';
+import Image from 'next/image';
+
+interface PositionTabProps {
+  showSubMenu: boolean;
+  subMenuClickHandler: () => void;
+  activePositionTabMobile: string;
+  openTabClickHandler: () => void;
+  historyTabClickHandler: () => void;
+}
 
 type TranslateFunction = (s: string) => string;
-const PositionTab = () => {
+
+const PositionTab = ({
+  showSubMenu,
+  subMenuClickHandler,
+  activePositionTabMobile,
+  openTabClickHandler,
+  historyTabClickHandler,
+}: PositionTabProps) => {
   const {t}: {t: TranslateFunction} = useTranslation('common');
-  const [activeTab, setActiveTab] = useState('Open');
+  const globalCtx = useGlobal();
+
+  // const [activeTab, setActiveTab] = useState(POSITION_TAB.OPEN);
 
   const tabBodyWidth = 'w-320px';
 
-  const openTabClickHandler = () => {
-    setActiveTab('Open');
-  };
+  // const openTabClickHandler = () => {
+  //   setActiveTab(POSITION_TAB.OPEN);
+  // };
 
-  const historyTabClickHandler = () => {
-    setActiveTab('History');
-  };
+  // const historyTabClickHandler = () => {
+  //   setActiveTab(POSITION_TAB.HISTORY);
+  // };
 
-  const currentSubTab = activeTab === 'Open' ? <OpenSubTab /> : <HistorySubTab />;
+  const currentSubTab =
+    globalCtx.layoutAssertion === LayoutAssertion.MOBILE ? (
+      activePositionTabMobile === POSITION_TAB.OPEN ? (
+        <OpenSubTab />
+      ) : (
+        <HistorySubTab />
+      )
+    ) : activePositionTabMobile === POSITION_TAB.OPEN ? (
+      <OpenSubTab />
+    ) : (
+      <HistorySubTab />
+    );
 
   const activeOpenTabStyle =
-    activeTab == 'Open' ? 'bg-darkGray8 text-lightWhite' : 'bg-darkGray6 text-lightGray';
+    activePositionTabMobile == POSITION_TAB.OPEN
+      ? 'bg-darkGray8 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
+
   const activeHistoryTabStyle =
-    activeTab == 'History' ? 'bg-darkGray8 text-lightWhite' : 'bg-darkGray6 text-lightGray';
+    activePositionTabMobile == POSITION_TAB.HISTORY
+      ? 'bg-darkGray8 text-lightWhite'
+      : 'bg-darkGray6 text-lightGray';
 
   const tabPart = (
     <>
@@ -47,32 +85,53 @@ const PositionTab = () => {
       </ul>
     </>
   );
-  return (
+
+  const subMenu = (
+    <div
+      className={`flex h-screen w-screen flex-col items-center justify-center bg-darkGray ${
+        showSubMenu ? 'visible translate-y-0 opacity-100' : 'invisible translate-y-full opacity-0'
+      } absolute left-0 ${'bottom-76px'} overflow-hidden pt-40 transition-all duration-150`}
+    >
+      <div className="mb-3 mr-30px flex self-end sm:pr-30px">
+        <ImCross onClick={subMenuClickHandler} className="cursor-pointer" />
+      </div>
+      {currentSubTab}
+    </div>
+  );
+
+  const desktopLayout = (
     <div
       className={`pointer-events-none fixed top-82px right-0 z-10 flex overflow-x-hidden overflow-y-hidden outline-none focus:outline-none`}
     >
       <div className="relative my-6 mx-auto w-auto max-w-xl">
         {' '}
         <div className={`relative`}>
-          {/* sidebar self */}
+          {/* Info: sidebar self (20230808 - Shirley) */}
           <div
             className={`pointer-events-auto ${tabBodyWidth} h-screen bg-darkGray p-5 text-white transition-all duration-300`}
           >
             {tabPart}
             {currentSubTab}
-            {/* <h1 className="pl-5 text-2xl font-bold">Order information</h1> */}
-
-            {/* <div className="mt-20">
-            <NotificationItem />
-          </div>
-          <div className="mt-5">
-            <NotificationItem />
-          </div> */}
           </div>
         </div>
       </div>
     </div>
   );
+
+  const mobileLayout = (
+    <div className="rounded-md bg-darkGray5 p-3">
+      <button onClick={subMenuClickHandler} className="m-auto block">
+        <Image src="/elements/position_tab_icon.svg" alt="position_icon" width={25} height={25} />
+      </button>
+
+      {subMenu}
+    </div>
+  );
+
+  const displayedLayout =
+    globalCtx.layoutAssertion === LayoutAssertion.MOBILE ? mobileLayout : desktopLayout;
+
+  return <>{displayedLayout}</>;
 };
 
 export default PositionTab;
