+++ src/components/trade_page_body/trade_page_body.tsx
@@ -1,7 +1,5 @@
 import OrderSection from '../order_section/order_section';
 import MarketSection from '../market_section/market_section';
-import {useGlobal} from '../../contexts/global_context';
-import OrderSectionMobile from '../order_section_mobile/order_section_mobile';
 import {IRecommendedNews} from '../../interfaces/tidebit_defi_background/news';
 
 interface ITradePageBodyProps {
@@ -9,10 +7,7 @@ interface ITradePageBodyProps {
 }
 
 const TradePageBody = (props: ITradePageBodyProps) => {
-  const {layoutAssertion} = useGlobal();
-
-  const displayedOrdersection =
-    layoutAssertion === 'mobile' ? <OrderSectionMobile /> : <OrderSection />;
+  const displayedOrdersection = <OrderSection />;
 
   return (
     <>
