+++ src/constants/config.ts
@@ -280,7 +280,7 @@ export const PRIVATE_POLICY = DOMAIN + '{hash}';
 export const DeWT_VALIDITY_PERIOD = 60 * 60; // seconds
 export const CFD_LIQUIDATION_TIME = 604800;
 export const TP_SL_LIMIT_RATIO = 0;
-export const DEFAULT_TICKER = 'ETH-USDT';
+export const DEFAULT_CRYPTO = 'ETH';
 export const DEFAULT_INSTID = 'ETH-USDT';
 export const DEFAULT_CURRENCY = Currency.ETH;
 export const SHARING_BG_IMG_THRESHOLD_PNL_PERCENT = 5;
