+++ src/constants/layout_assertion.ts
@@ -1,0 +1,9 @@
+export type ILayoutAssertion = 'MOBILE' | 'DESKTOP';
+export interface ILayoutAssertionConstant {
+  MOBILE: ILayoutAssertion;
+  DESKTOP: ILayoutAssertion;
+}
+export const LayoutAssertion: ILayoutAssertionConstant = {
+  MOBILE: 'MOBILE',
+  DESKTOP: 'DESKTOP',
+};
