+++ src/pages/500.tsx
@@ -2,7 +2,6 @@ import React, {useContext, useEffect} from 'react';
 import Lottie from 'lottie-react';
 import errorAnimation from '../../public/animation/oops.json';
 import NavBar from '../components/nav_bar/nav_bar';
-import NavBarMobile from '../components/nav_bar_mobile/nav_bar_mobile';
 import {serverSideTranslations} from 'next-i18next/serverSideTranslations';
 import {AppContext} from '../contexts/app_context';
 import {useGlobal} from '../contexts/global_context';
@@ -11,6 +10,7 @@ import Footer from '../components/footer/footer';
 import Head from 'next/head';
 import {TIDEBIT_FAVICON} from '../constants/display';
 import {useTranslation} from 'next-i18next';
+import {LayoutAssertion} from '../constants/layout_assertion';
 
 type TranslateFunction = (s: string) => string;
 
@@ -18,7 +18,8 @@ const Custom500 = () => {
   const {t}: {t: TranslateFunction} = useTranslation('common');
 
   const {layoutAssertion} = useGlobal();
-  const displayedNavBar = layoutAssertion === 'mobile' ? <NavBarMobile /> : <NavBar />;
+  const displayedNavBar = <NavBar />;
+
   const appCtx = useContext(AppContext);
 
   useEffect(() => {
