+++ src/pages/leaderboard.tsx
@@ -1,6 +1,5 @@
 import React, {useContext, useEffect} from 'react';
 import NavBar from '../components/nav_bar/nav_bar';
-import NavBarMobile from '../components/nav_bar_mobile/nav_bar_mobile';
 import Head from 'next/head';
 import BoardPageBody from '../components/board_page_body/board_page_body';
 import {serverSideTranslations} from 'next-i18next/serverSideTranslations';
@@ -8,10 +7,11 @@ import {AppContext} from '../contexts/app_context';
 import {useGlobal} from '../contexts/global_context';
 import {ILocale} from '../interfaces/tidebit_defi_background/locale';
 import {TIDEBIT_FAVICON} from '../constants/display';
+import {LayoutAssertion} from '../constants/layout_assertion';
 
 const Leaderboard = () => {
   const {layoutAssertion} = useGlobal();
-  const displayedNavBar = layoutAssertion === 'mobile' ? <NavBarMobile /> : <NavBar />;
+  const displayedNavBar = <NavBar />;
 
   const appCtx = useContext(AppContext);
